Replaced Long to Integer when counting monsters.

+++ src/main/java/monsterwars/main/InputArgumentValidator.java
@@ -5,10 +5,10 @@ class InputArgumentValidator {
     private static final String NOT_A_NUMBER_ERROR_MESSAGE = "Not a valid number as a parameter: ";
     private static final String NOT_ENOGH_MONSTERS_ERROR_MESSAGE = "Too few monsters are specified: ";
 
-    static Long validate(String[] args) {
-        Long numberOfMonsters = null;
+    static Integer validate(String[] args) {
+        Integer numberOfMonsters = null;
         try {
-            numberOfMonsters = Long.valueOf(args[0]);
+            numberOfMonsters = Integer.valueOf(args[0]);
         } catch (NumberFormatException exception) {
             System.out.println(NOT_A_NUMBER_ERROR_MESSAGE + args[0]);
         }
